Changed notional to Long, started implementing cucumber test

+++ src/main/scala/com/prystupa/MatchingAlgorithm.scala
@@ -43,7 +43,7 @@ trait MatchingAlgorithm {
 	 * @param notional - notional of order
 	 * @return - returns constructed matching unit
 	 */
-	def createMatch(order1: Order, order2: Order, notional: Option[Int]): MatchingUnit
+	def createMatch(order1: Order, order2: Order, notional: Option[Long]): MatchingUnit
 
 
 	/**
@@ -53,6 +53,6 @@ trait MatchingAlgorithm {
 	 * @param notional - notional of order
 	 * @return - returns constructed matching unit
 	 */
-	def createSwap(order1: Order, order2: Order, notional: Option[Int]): MatchingUnit
+	def createSwap(order1: Order, order2: Order, notional: Option[Long]): MatchingUnit
 
 }
