+++ src/main/scala/com/prystupa/MatchingUnit.scala
@@ -10,7 +10,7 @@ package com.prystupa
  * Date: 10/7/12
  * Time: 1:44 PM
  */
-class MatchingUnit(val notional: Int, val order1: Order, val order2: Order) {
+class MatchingUnit(val notional: Long, val order1: Order, val order2: Order) {
 
 	def flip():MatchingUnit = {
 		new MatchingUnit(notional, order2, order1)
